feat: Rename source file as requested
