+++ src/databricks/labs/ucx/source_code/languages.py
@@ -4,7 +4,7 @@ from databricks.labs.ucx.hive_metastore.table_migrate import MigrationIndex
 from databricks.labs.ucx.source_code.base import Fixer, Linter, SequentialLinter
 from databricks.labs.ucx.source_code.pyspark import SparkSql
 from databricks.labs.ucx.source_code.queries import FromTable
-from databricks.labs.ucx.source_code.dbfsvisitor import DBFSFinderLinter
+from databricks.labs.ucx.source_code.dbfs import DBFSFinderLinter
 
 
 class Languages:
